impl(server): Phase 2A r2 commit 4 — folder moves to protocol-aligned layers (#542)
Reshape `packages/server/src/` from concern-flavored folders to
protocol-aligned layers per plan v5 §2 mapping table:

- `auth/` → `identity/services/`.
- `services/auth.service.*`, `services/contact.service.*`,
  `services/participant.service.*`, `services/agent-visibility.*`,
  `services/session-validator.*` → `identity/services/`.
- `services/conversation.service.*`, `services/message.service.*`,
  `services/task.service.*`, `services/default-tm.*`,
  `services/conversation-admin-authority.*` → `task/services/`.
- `services/presence.service.*`, `services/presence-event-sink.*`
  → `network/services/`.
- `ws/connection.*` → `transport/connection.*`.
- `rpc/{context,define-layered-method,layer-scopes,layer-tags,layer-boundary.types-check}.ts`
  → `transport/`.

Kernels (`db/`, `crypto/`, `runtime/`, `runtime-surface/`, `adapters/`,
`config/`, `test-utils/`, `logger.ts`) stay at top level — they are
infrastructure rather than protocol layers (user-authorized
large-folder lint exemption per plan §10).

Import paths updated to reflect new depth. Cross-layer imports use
relative paths (`../../<layer>/foo.js`); commit 5 converts those to
workspace-name imports per plan §1.

`docs/server/rpc-handlers.mdx` updated: factory-name list replaced
with registry-name list reflecting the post-DI shape.

R6 — folder-path drift: `docs/architecture.mdx` and
`docs/development/contributing.mdx` carry no live references to the
moved paths. Historic plan docs under `docs/plans/` and
`docs/superpowers/` intentionally retain their original paths (they
document past work as of plan time).

R9 grep gate: `git grep 'as Effect.Effect<' packages/server/src/transport/context.ts`
returns zero hits. The gate's path tracks the context.ts move from
`rpc/` to `transport/`.
