+++ packages/server/src/app/layers.test.ts
@@ -6,15 +6,15 @@ import type { Db } from "../db/client.js";
 import type { Database } from "../db/database.js";
 import { NoopTraceCaptureLive } from "../runtime-surface/trace-capture.js";
 import { LoggerLive, getLogger } from "../logger.js";
-import { ConnectionManager } from "../ws/connection.js";
+import { ConnectionManager } from "../transport/connection.js";
 import { AgentEndpointResolver } from "../network/agent-endpoint-resolver.js";
 import { NetworkSendService } from "../network/network-send.js";
-import { AuthService } from "../services/auth.service.js";
-import { ContactsService } from "../services/contact.service.js";
-import { ConversationService } from "../services/conversation.service.js";
-import { MessageService } from "../services/message.service.js";
-import { ParticipantService } from "../services/participant.service.js";
-import { PresenceService } from "../services/presence.service.js";
+import { AuthService } from "../identity/services/auth.service.js";
+import { ContactsService } from "../identity/services/contact.service.js";
+import { ConversationService } from "../task/services/conversation.service.js";
+import { MessageService } from "../task/services/message.service.js";
+import { ParticipantService } from "../identity/services/participant.service.js";
+import { PresenceService } from "../network/services/presence.service.js";
 import { AppHost } from "./app-host.js";
 import { WebhookClient } from "../adapters/webhook.js";
 import {
