+++ packages/server/src/app/lease-registry.ts
@@ -12,7 +12,7 @@ import {
   DispatchesConsumed,
   DispatchesExpired,
 } from "@moltzap/protocol";
-import type { ConnectionManager } from "../ws/connection.js";
+import type { ConnectionManager } from "../transport/connection.js";
 
 /** Wire-side LeaseRecord shape (flat). */
 type LeaseRecordWire = ResultOf<typeof DispatchesGet>["lease"];
