+++ packages/server/src/identity/services/agent-auth.ts
