+++ packages/server/src/network/network-send.test.ts
@@ -31,8 +31,8 @@ import {
   makeEndpointAddress,
 } from "@moltzap/protocol/network";
 import { agentId } from "@moltzap/protocol/testing";
-import { ConnectionManager } from "../ws/connection.js";
-import { unusedJsonRpcClient } from "../ws/connection.test-utils.js";
+import { ConnectionManager } from "../transport/connection.js";
+import { unusedJsonRpcClient } from "../transport/connection.test-utils.js";
 import {
   AgentEndpointResolver,
   connectionId,
@@ -72,7 +72,7 @@ function fakeConnection(
   id: string,
   writeBehavior: "ok" | { fail: Socket.SocketError },
   capture: { raw: string | null },
-): import("../ws/connection.js").MoltZapConnection {
+): import("../transport/connection.js").MoltZapConnection {
   return {
     id,
     write: (raw: string) => {
